issue #608 move Marker, ProblemMarker interfaces under "common"
