+++ packages/markers/src/browser/problem/problem-frontend-module.ts
@@ -10,7 +10,8 @@ import { ProblemWidget } from './problem-widget';
 import { ProblemContribution } from './problem-contribution';
 import { createProblemWidget } from './problem-container';
 import { CommandContribution, MenuContribution, KeybindingContribution } from "@theia/core/lib/common";
-import { ProblemManager, PROBLEM_KIND } from './problem-marker';
+import { ProblemManager } from './problem-marker';
+import { PROBLEM_KIND } from '../../common/problem-marker';
 import { WidgetFactory } from '@theia/core/lib/browser/widget-manager';
 
 import '../../../src/browser/style/index.css';
