+++ packages/markers/src/browser/problem/problem-marker.ts
@@ -5,23 +5,13 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-import { Marker, MarkerManager } from '../marker-manager';
+import { MarkerManager } from '../marker-manager';
 import { Diagnostic } from "vscode-languageserver-types";
 import { injectable, inject } from 'inversify';
 import { StorageService } from '@theia/core/lib/browser/storage-service';
 import { FileSystemWatcher } from '@theia/filesystem/lib/common';
+import { PROBLEM_KIND } from '../../common/problem-marker';
 
-export const PROBLEM_KIND = 'problem';
-
-export interface ProblemMarker extends Marker<Diagnostic> {
-    kind: 'problem';
-}
-
-export namespace ProblemMarker {
-    export function is(node: Marker<object>): node is ProblemMarker {
-        return 'kind' in node && node.kind === PROBLEM_KIND;
-    }
-}
 
 @injectable()
 export class ProblemManager extends MarkerManager<Diagnostic> {
